Fix int vs long int mismatches in printf args.
2007-01-18  Behdad Esfahbod  <behdad@gnome.org>

        * src/table.c (main):
        * src/trie.c (_vte_trie_addx), (main):
        * src/utf8echo.c (main):
        Fix int vs long int mismatches in printf args.


svn path=/trunk/; revision=1453

+++ src/table.c
@@ -851,7 +851,7 @@ main(int argc, char **argv)
 		printf("`%s' => `%s'", tmp, (result ? result : "(NULL)"));
 		g_free(tmp);
 		print_array(array);
-		printf(" (%d chars)\n", consumed ? consumed - candidate: 0);
+		printf(" (%d chars)\n", (int) (consumed ? consumed - candidate: 0));
 		g_free(candidate);
 	}
 	_vte_table_free(table);
